第一次提交

+++ src/main/java/com/ipmapping/cnIP/IPCalculate.java
@@ -7,7 +7,7 @@ package com.ipmapping.cnIP;
  */
 public class IPCalculate {
 
-    public static int validateIP(String ip) {
+    public static Long validateIP(String ip) {
 
         String[] ipArray = ip.split("\\.");
         int[] ipNum = new int[ipArray.length];
@@ -16,7 +16,7 @@ public class IPCalculate {
         }
 
         // 各个数字乘以相应的数量级
-        int ipNumTotal = ipNum[0] * 256 * 256 * 256 + ipNum[1] * 256 * 256 + ipNum[2] * 256 + ipNum[3];
+        long ipNumTotal = (long)ipNum[0] * 256 * 256 * 256 + ipNum[1] * 256 * 256 + ipNum[2] * 256 + ipNum[3];
 
         return ipNumTotal;
     }
